Change default scheme for provider.
This is a possibly breaking change as it changes the resolution order
of how we decide what scheme to use when we connect to the provider.

Earlier we picked http if no scheme was provided in the Vagrantfile,
thus overriding any setting in the .fog file or the defaults in the provider.

With this change, we can specify a scheme in the Vagrantfile, the .fog file or
use the providers default as a fallback.